refactor: extract OpenAIEndpointBuilder.swift from OpenAIErrorMapper.swift
Extracts the `OpenAIEndpointBuilder` enum (14 lines) from
OpenAIErrorMapper.swift into its own file. OpenAIErrorMapper.swift is
kept as the home for the error-mapping enum (matching its filename).

Byte-identical move of lines 3-16. Only imports Foundation.

OpenAIErrorMapper.swift: 128 → 113 lines. Tests: 667.
