+++ Sources/BugNarrator/Services/OpenAIErrorMapper.swift
@@ -1,20 +1,5 @@
 import Foundation
 
-enum OpenAIEndpointBuilder {
-    static func endpoint(for path: String, baseURL: URL) -> URL {
-        var pathComponents = path.split(separator: "/").map(String.init)
-
-        if let firstComponent = pathComponents.first,
-           baseURL.lastPathComponent.caseInsensitiveCompare(firstComponent) == .orderedSame {
-            pathComponents.removeFirst()
-        }
-
-        return pathComponents.reduce(baseURL) { url, component in
-            url.appendingPathComponent(component)
-        }
-    }
-}
-
 enum OpenAIErrorMapper {
     static func mapResponse(
         statusCode: Int,
